+++ qt4-gui/src/config/skin.h
@@ -39,17 +39,17 @@ namespace Config
 class Border
 {
 public:
-  unsigned short top, bottom;
-  unsigned short left, right;
+  int top, bottom;
+  int left, right;
 
-  void AdjustForMenuBar(unsigned short h1, unsigned short h2);
+  void AdjustForMenuBar(int h1, int h2);
 };
 
 class FrameSkin
 {
 public:
   struct Border border;
-  unsigned short frameStyle;
+  unsigned frameStyle;
   bool maintainBorder;
   bool hasMenuBar;
   bool transparent;
@@ -70,7 +70,7 @@ public:
   virtual ~ShapeSkin() { }
   virtual void loadSkin(CIniFile& skinFile, const QString& name);
   QRect borderToRect(const QWidget* w) const;
-  void AdjustForMenuBar(unsigned short h1, unsigned short h2);
+  void AdjustForMenuBar(int h1, int h2);
 };
 
 class ButtonSkin : public ShapeSkin
@@ -92,9 +92,9 @@ class LabelSkin : public ShapeSkin
 {
 public:
   QPixmap pixmap;
-  unsigned short frameStyle;
+  unsigned frameStyle;
   bool transparent;
-  unsigned short margin;
+  int margin;
 
   virtual ~LabelSkin() { }
   virtual void loadSkin(CIniFile& skinFile, const QString& name, const QString& baseSkinDir);
@@ -144,7 +144,7 @@ public:
   QPixmap mainwinMask(int width, int height) const;
 
   void setFrameTransparent(bool transparent);
-  void setFrameStyle(unsigned short frameStyle);
+  void setFrameStyle(unsigned frameStyle);
 
   FrameSkin frame;
   ButtonSkin btnSys;
@@ -176,7 +176,7 @@ public:
   bool tileGroupBackImage;
 
   // Functions
-  void AdjustForMenuBar(unsigned short n);
+  void AdjustForMenuBar(int n);
   int frameWidth(void);
   int frameHeight(void);
   QPalette palette(QWidget* parent);
@@ -193,7 +193,7 @@ private:
   QPixmap scaleWithBorder(const QPixmap& pm, int width, int height) const;
 
   QString mySkinName;
-  unsigned short myMenuBarHeight;
+  int myMenuBarHeight;
 };
 
 } // namespace Config
